Phase 2 step 10: Migrate leaf mixins: HeaderStatusMixin, LiveTMSMixin, OrderBookMixin
